refactor(View): move View implementation to View directory
