+++ src/model/screens/mod.rs
@@ -1,29 +1,5 @@
-use crate::viewmodels::ViewModelState;
-
 pub mod bookmarked;
 pub mod details_actions;
 pub mod edit_config;
 pub mod latest;
 pub mod mail_list;
-
-#[derive(Copy, Debug, Clone, PartialEq, Eq, Hash)]
-pub enum View {
-    MailingListSelection,
-    BookmarkedPatchsets,
-    LatestPatchsets,
-    PatchsetDetails,
-    EditConfig,
-}
-
-impl View {
-    #[allow(dead_code, unused_variables)]
-    pub fn draw_screen(&self, state: ViewModelState) {
-        match self {
-            View::MailingListSelection => todo!(),
-            View::BookmarkedPatchsets => todo!(),
-            View::LatestPatchsets => todo!(),
-            View::PatchsetDetails => todo!(),
-            View::EditConfig => todo!(),
-        }
-    }
-}
